Move feature flag utility function into superset/__init__.py
